creation of clidDisplay file to present the viewer with the required details
